+++ packages/mui/src/icons.tsx
@@ -1,13 +1,20 @@
-import type { BuilderActionIconType } from "@form-engine-ts/react";
+import type { BuilderActionIconType, QuestionType } from "@form-engine-ts/react";
 import {
   Add,
   ArrowDownward,
+  ArrowDropDown,
   ArrowUpward,
+  CheckBox,
   Close,
   Delete,
   DragHandle,
   Edit,
+  Numbers,
+  RadioButtonChecked,
   Settings,
+  Star,
+  Subject,
+  TextFields,
   Translate
 } from "@mui/icons-material";
 import type { ReactNode } from "react";
@@ -32,5 +39,26 @@ export function muiDefaultIconResolver(actionType: BuilderActionIconType): React
       return <Close fontSize="small" />;
     case "dragHandle":
       return <DragHandle fontSize="small" />;
+  }
+}
+
+export function muiDefaultFieldTypeIcon(type: QuestionType): ReactNode {
+  switch (type) {
+    case "text":
+      return <TextFields fontSize="small" />;
+    case "textarea":
+      return <Subject fontSize="small" />;
+    case "number":
+      return <Numbers fontSize="small" />;
+    case "rating":
+      return <Star fontSize="small" />;
+    case "select":
+      return <ArrowDropDown fontSize="small" />;
+    case "multi-select":
+      return <CheckBox fontSize="small" />;
+    case "checkbox":
+      return <CheckBox fontSize="small" />;
+    case "radio":
+      return <RadioButtonChecked fontSize="small" />;
   }
 }